backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
arxiv
chromadb
langgraph
reportlab
PyPDF2
aiofiles
fitz
pytesseract
